+++ src/errors/RequestCancelledError.ts
@@ -1,7 +1,9 @@
 export const cancelledRequestStatus = 499
 
+export const cancelledErrorCode = 'request_cancelled'
+
 export class RequestCancelledError extends Error {
-  public code = 'request_cancelled'
+  public code = cancelledErrorCode
 
   constructor(message: string) {
     super(message)