+++ src/service/http/middlewares/requestStats.ts
@@ -25,21 +25,21 @@ class IncomingRequestStats {
 
 export const incomingRequestStats = new IncomingRequestStats()
 
-const requestClosed = <T extends IOClients, U, V>(ctx: ServiceContext<T, U, V>) => () => {
+const requestClosed = () => {
   incomingRequestStats.closed++
+}
+const requestAborted = <T extends IOClients, U, V>(ctx: ServiceContext<T, U, V>) => () => {
+  incomingRequestStats.aborted++
 
   if (ctx.vtex.cancellation && ctx.vtex.cancellation.cancelable) {
     ctx.vtex.cancellation.source.cancel(cancelMessage)
     ctx.vtex.cancellation.cancelled = true
   }
 }
-const requestAborted = () => {
-  incomingRequestStats.aborted++
-}
 
 export async function trackIncomingRequestStats <T extends IOClients, U, V> (ctx: ServiceContext<T, U, V>, next: () => Promise<any>) {
-  ctx.req.on('close', requestClosed(ctx))
-  ctx.req.on('aborted', requestAborted)
+  ctx.req.on('close', requestClosed)
+  ctx.req.on('aborted', requestAborted(ctx))
   incomingRequestStats.total++
   await next()
 }